#956 Phase 2: extract cos/flow_hash.rs from tx.rs
Continues the cos/ submodule decomposition started by Phase 1
(cos/ecn.rs at PR #976).

Moved from tx.rs to cos/flow_hash.rs (~150 LOC, 6 functions):
- mix_cos_flow_bucket: file-private XorShift mix step
- exact_cos_flow_bucket: file-private 5-tuple → bucket hash
- cos_flow_hash_seed_from_os: per-queue salt drawn from getrandom
  (with CLOCK_MONOTONIC + pid + stack-addr fallback)
- cos_item_flow_key: CoSPendingTxItem → SessionKey accessor
- cos_flow_bucket_index: hash → masked bucket index for SFQ
- cos_queue_prospective_active_flows: admission gate denominator

Production callers re-exported from cos/mod.rs for the short
super::cos::{...} import path on tx.rs sites at 4060/4074/4119,
4138/4188/4305/4326/5987, 4283/4317/4624, and 5830 (verified by
Codex round-2 grep walk).

mix_cos_flow_bucket and exact_cos_flow_bucket stay file-private
to flow_hash.rs — only used internally; tests reach
cos_flow_bucket_index (production-exposed) instead.

Constants COS_FLOW_FAIR_BUCKETS / COS_FLOW_FAIR_BUCKET_MASK
stay in afxdp::types (their original home; they size other
types there). flow_hash.rs imports the mask.

Plan iteration: 3 rounds of Codex hostile review before
PLAN-READY at round 3 (caught 5+5 factual corrections —
constant ownership, false worker.rs claim, over-exposed file-
private helpers, incomplete production-call-site list, wrong
SessionKey path). Gemini adversarial daemon dropped state on
its review attempt; proceeded on Codex's PLAN-READY since the
refactor is bounded and follows Phase 1's already-cleared
pattern.

cargo test --release: 865 passed (no count change), 0 failed.
Build clean.
